requirements.txt
streamlit
pandas
openpyxl
xlrd
requests
python-dotenv